+++ src/main/java/com/viettridao/cafe/dto/response/menu_item/MenuItemResponse.java
@@ -1,5 +1,6 @@
 package com.viettridao.cafe.dto.response.menu_item;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -12,7 +13,7 @@ import lombok.Setter;
 public class MenuItemResponse {
     private Integer id;
     private String itemName;
-    private Double currentPrice;
+    private BigDecimal currentPrice;
     private Boolean isDeleted;
     private List<MenuDetailResponse> menuDetails;
 }
